demo/apps/aks-to-aws/requirements.txt
boto3==1.34.34
azure-identity==1.15.0
azure-core==1.29.7